clients/cli-chatbot/pyproject.toml
[project]
name = "cli-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "mcp[cli]>=1.9.4",
    "ollama>=0.5.1",
    "python-dotenv>=1.1.0",
]
